fix update cart + add api update+delete (by id)

+++ ProductService/src/main/java/com/thc/productservice/Repository/Impl/ProductRepositoryImpl.java
@@ -30,7 +30,7 @@ public class ProductRepositoryImpl implements ProductRepository {
     }
 
     @Override
-    public int count() {
+    public Long count() {
         return productMapper.count();
     }
